Refactor PHPWord to support PHP 8 type declarations
This commit updates several files, primarily within the 'src/PhpWord/Element' directory, to use the new type declaration syntax introduced in PHP 8. Variables that were previously declared using a comment block above the variable declaration are now using actual type declarations. This change helps the static analysis tool to better understand the types of variables, ensures that the value assigned to a variable is of the correct type, and, thus, improves maintainability and readability of the code.

In addition, the TestableTemplateProcesor class was renamed to TestableTemplateProcessor, fixing a typo. The composer.json file was also reformatted to make it easier to read, but no substantial changes were made to it.

Tests are modified to reflect the new class name and additional type declarations. The changes made aim to improve code quality and prevent potential type-related errors in the future.