Update handling of blocks and text elements in PHPWord
Changes are made to PHPWord writer, abstract container, shared HTML, and template processor. The aim is to improve and correct how blocks of text and individual elements within them are processed.

In the TOC writer, checks for TextRun elements are added. The contents of these elements are now correctly extracted and written. The handling of ListItemRun in AbstractContainer now includes TextRun as possible parent, which helps handle more diverse document structures.

In Shared/Html, the case of 'margin-left' is added and the value now converted to points. TemplateProcessor's methods cloneBlock and replaceBlock are significantly modified. Cloning no longer requires replacement of the original block and can index variables. Block replacement has increased checks and conditions for better accuracy. These changes improve the output quality and flexibility of TemplateProcessor.

Overall, the codebase shows improved flexibility and accuracy in handling text elements and blocks within documents.